test: clear tasks umbrella MEDIUM review items
