+++ test/features/tasks/TEST_REVIEW.md
@@ -56,15 +56,20 @@ Generated by automated test-quality review (static analysis); aggregates the per
 
 - [x] **[HIGH]** Add `test/features/tasks/ui/model/task_browse_row_interactions_test.dart` — 63-line pure-function file with no tests at all; excellent Glados candidate (`buildTaskBrowseRowInteraction` invariants, `taskRowInteractionPriority` priority rules). Referenced: [ui/model/TEST_REVIEW.md](ui/model/TEST_REVIEW.md).
 
-- [ ] **[MED]** Merge `task_filter_modal_save_flow_test.dart` into `task_filter_modal_test.dart` to fix the one-test-file-per-source violation; consolidate the 3 near-identical `buildSubject()` / `buildWithSaveEnabled()` / `buildWithProjects()` factory functions into a single `_buildSubject({...})` helper. Also fix 2 remaining ad-hoc `getIt` setups in `task_category_filter_test.dart` and `task_label_quick_filter_test.dart`. Referenced: [ui/filtering/TEST_REVIEW.md](ui/filtering/TEST_REVIEW.md).
+- [x] **[MED]** Merge `task_filter_modal_save_flow_test.dart` into `task_filter_modal_test.dart` to fix the one-test-file-per-source violation; consolidate the 3 near-identical `buildSubject()` / `buildWithSaveEnabled()` / `buildWithProjects()` factory functions into a single `_buildSubject({...})` helper. Also fix 2 remaining ad-hoc `getIt` setups in `task_category_filter_test.dart` and `task_label_quick_filter_test.dart`. Referenced: [ui/filtering/TEST_REVIEW.md](ui/filtering/TEST_REVIEW.md).
+  - **RESOLVED:** (stale) — all referenced sub-items are resolved in ui/filtering: the save-flow file was merged back, the three factories are consolidated onto `buildSubject`, and both ad-hoc getIt setups were converted to `setUpTestGetIt`.
 
-- [ ] **[MED]** Add Glados property tests: `ChecklistController._insertItemAt` (list-move idempotency, index clamping), `SavedTaskFilter` JSON round-trip (full enum/set field coverage), `sumTimeSpentFromEntities` (union-interval invariants — result never negative, never exceeds naive sum). Referenced: [state/TEST_REVIEW.md](state/TEST_REVIEW.md), [repository/TEST_REVIEW.md](repository/TEST_REVIEW.md).
+- [x] **[MED]** Add Glados property tests: `ChecklistController._insertItemAt` (list-move idempotency, index clamping), `SavedTaskFilter` JSON round-trip (full enum/set field coverage), `sumTimeSpentFromEntities` (union-interval invariants — result never negative, never exceeds naive sum). Referenced: [state/TEST_REVIEW.md](state/TEST_REVIEW.md), [repository/TEST_REVIEW.md](repository/TEST_REVIEW.md).
+  - **RESOLVED:** (stale) — all three Glados families exist: `debugInsertItemAt` property in checklist_controller_test, the SavedTaskFilter fromJson/toJson round-trip property in the saved-filters tests, and the `sumTimeSpentFromEntities`/`buildTimeRanges` property groups in task_progress_repository_test.
 
-- [ ] **[MED]** Fix one-test-file-per-source violations in `ui/` group: consolidate `title_text_field_shortcuts_test.dart` + `title_text_field_tap_outside_test.dart` into `title_text_field_test.dart`; merge orphan timer-text test files (`compact_task_progress_timer_text_test.dart`, `linked_duration_timer_text_test.dart`) into their primary test files. Referenced: [ui/TEST_REVIEW.md](ui/TEST_REVIEW.md).
+- [x] **[MED]** Fix one-test-file-per-source violations in `ui/` group: consolidate `title_text_field_shortcuts_test.dart` + `title_text_field_tap_outside_test.dart` into `title_text_field_test.dart`; merge orphan timer-text test files (`compact_task_progress_timer_text_test.dart`, `linked_duration_timer_text_test.dart`) into their primary test files. Referenced: [ui/TEST_REVIEW.md](ui/TEST_REVIEW.md).
+  - **RESOLVED:** done — the title_text_field files were merged earlier; the two orphan timer-text files (`compact_task_progress_timer_text_test.dart`, `linked_duration_timer_text_test.dart`) are now merged into their primary mirrors and deleted.
 
-- [ ] **[MED]** Replace `Future.delayed(Duration.zero)` in `checklist_controller_test.dart` (lines 910–911, 952) and real `.timeout(Duration(seconds:1))` waits in `task_progress_controller_test.dart` with `fakeAsync` + `async.flushMicrotasks()` — these are fake-time policy violations that can mask flakiness. Referenced: [state/TEST_REVIEW.md](state/TEST_REVIEW.md).
+- [x] **[MED]** Replace `Future.delayed(Duration.zero)` in `checklist_controller_test.dart` (lines 910–911, 952) and real `.timeout(Duration(seconds:1))` waits in `task_progress_controller_test.dart` with `fakeAsync` + `async.flushMicrotasks()` — these are fake-time policy violations that can mask flakiness. Referenced: [state/TEST_REVIEW.md](state/TEST_REVIEW.md).
+  - **RESOLVED:** (stale) — no `Future.delayed(Duration.zero)` remains in checklist_controller_test and no real `.timeout(...)` waits remain in task_progress_controller_test.
 
-- [ ] **[MED]** Add `ChecklistCompletionRateController` tests (completely untested class in `checklist_controller.dart`): rate as `completedCount / totalCount`, division-by-zero guard, reactive update. Referenced: [state/TEST_REVIEW.md](state/TEST_REVIEW.md).
+- [x] **[MED]** Add `ChecklistCompletionRateController` tests (completely untested class in `checklist_controller.dart`): rate as `completedCount / totalCount`, division-by-zero guard, reactive update. Referenced: [state/TEST_REVIEW.md](state/TEST_REVIEW.md).
+  - **RESOLVED:** (stale) — `ChecklistCompletionRateController` has a dedicated group in `checklist_completion_controller_test.dart` (it lives in `checklist_completion_controller.dart`, not `checklist_controller.dart`).
 
 - [ ] **[LOW]** Centralize the 4+ inline `MockNavService` / `MockEntitiesCacheService` / `MockTimeService` / `MockTask` definitions that appear in `task_status_test.dart`, `task_action_bar_test.dart`, `task_status_modal_content_test.dart`, and list_cards test files — add each to `test/mocks/mocks.dart` and remove the inline definitions. Referenced: [ui/TEST_REVIEW.md](ui/TEST_REVIEW.md), [ui/header/TEST_REVIEW.md](ui/header/TEST_REVIEW.md), [ui/widgets/TEST_REVIEW.md](ui/widgets/TEST_REVIEW.md).
